chore(useMounted): deprecate useMounted hook since it does not work unless the useEffect is "last" in the render cycle
