+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
 Prepend new items to make git merging easier.
+- chore(useMounted): deprecate useMounted hook since it does not work unless the useEffect is "last" in the render cycle
 - fix(multi-select cell): add validation message as a tooltip on the invalid icon
 - fix(multi-autocompleter): do not call onChange when loading selected values on first-time render
 - fix(multi-choice control): always load provided selected values in case search pagination does not return a result set with the selected values
