+++ src/components/control/autocompleter.jsx
@@ -1,7 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react';
 import useFetch from '@bloodyaugust/use-fetch';
-import useMounted from '../../hooks/use-mounted.js';
 import Select from './select.jsx';
 import ListOption from '../list-option/list-option.jsx';
 import { API_ROOT } from '../../mocks/mock-constants.js';
@@ -9,7 +8,7 @@ import { API_ROOT } from '../../mocks/mock-constants.js';
 const Autocompleter = forwardRef(function Autocompleter(props, ref) {
   const { execute: executeModels } = useFetch();
   const { execute: executeValue } = useFetch();
-  const mounted = useMounted();
+  const mounted = useRef(false);
   const [models, setModels] = useState([]);
   const [model, setModel] = useState({ id: props.value });
   const selectRef = useRef();
@@ -68,6 +67,10 @@ const Autocompleter = forwardRef(function Autocompleter(props, ref) {
   }
 
   const listOptionRefs = models.map(() => React.createRef());
+
+  useEffect(() => {
+    mounted.current = true;
+  }, []);
 
   useImperativeHandle(ref, () => ({
     ...selectRef.current,
