+++ src/components/control/input.jsx
@@ -1,9 +1,8 @@
 import PropTypes from 'prop-types';
-import React, { forwardRef, useEffect, useImperativeHandle, useRef } from 'react';
+import React, { forwardRef, useEffect, useImperativeHandle, useLayoutEffect, useRef } from 'react';
 import cautionSvg from '../../svgs/caution.svg';
 import Icon from '../icon/icon.jsx';
 import styles from './input.css';
-import useMounted from '../../hooks/use-mounted.js';
 import useForwardedRef from '../../hooks/use-forwarded-ref.js';
 import useValidation from '../../hooks/use-validation.jsx';
 
@@ -24,7 +23,6 @@ const Input = forwardRef(function Input(props, forwardedRef) {
   };
 
   const ref = useForwardedRef(forwardedRef);
-  const mounted = useMounted();
   const [validationMessage, validate] = useValidation(props.validationMessage, refs.input);
 
   function onBlur(event) {
@@ -32,9 +30,7 @@ const Input = forwardRef(function Input(props, forwardedRef) {
     props.onBlur(event);
   }
 
-  useEffect(() => { // TODO: Convert to useLayoutEffect
-    if (!mounted.current) return;
-
+  useLayoutEffect(() => {
     // The MDS Input is using an uncontrolled `<input>`.
     // In order to set a new provided value prop, we
     // set the internal state of the `<input>`.
