+++ src/components/popover/popover.jsx
@@ -10,7 +10,7 @@ import IconButton from '../icon-button/icon-button.jsx';
 import iconClear from '../../svgs/clear.svg';
 import styles from './popover.css';
 import useFlush from '../../hooks/use-flush.js';
-import useMounted from '../../hooks/use-mounted.js';
+import useMountedEffect from '../../hooks/use-mounted-effect.js';
 import useForwardedRef from '../../hooks/use-forwarded-ref.js';
 
 const Popover = forwardRef(function Popover(props, ref) {
@@ -19,7 +19,6 @@ const Popover = forwardRef(function Popover(props, ref) {
   const sectionRef = useRef();
   const selfRef = useForwardedRef(ref);
   const { flush } = useFlush({ ref: sectionRef, initialDirection: props.flush, autoflush: props.autoflush, open });
-  const mounted = useMounted();
 
   function onBlur(event) {
     // Target is set to `null` when losing focus to a non-interactive element (e.g. text)
@@ -30,11 +29,13 @@ const Popover = forwardRef(function Popover(props, ref) {
   }
 
   useLayoutEffect(() => {
-    if (!mounted.current) return;
-
     if (open) {
       closeIconRef.current.focus({ preventScroll: true });
-    } else {
+    }
+  }, [open]);
+
+  useMountedEffect(() => {
+    if (!open) {
       props.onClose();
     }
   }, [open]);
